+++ collectors/lib/hadoop_http.py
@@ -85,11 +85,11 @@ class HadoopHttp(object):
 
     def emit_metric(self, context, current_time, metric_name, value, tag_dict=None):
         if not tag_dict:
-            print("%s.%s.%s.%s %d %d" % (self.service, self.daemon, ".".join(context), metric_name, current_time, value))
+            print("%s.%s.%s.%s %d %d" % (self.service, self.daemon, ".".join(context), metric_name, current_time, float(value)))
         else:
             tag_string = " ".join([k + "=" + v for k, v in tag_dict.iteritems()])
             print ("%s.%s.%s.%s %d %d %s" % \
-                  (self.service, self.daemon, ".".join(context), metric_name, current_time, value, tag_string))
+                  (self.service, self.daemon, ".".join(context), metric_name, current_time, float(value), tag_string))
         # flush to protect against subclassed collectors that output few metrics not having enough output to trigger
         # buffer flush within 10 mins, which then get killed by TCollector due to "inactivity"
         sys.stdout.flush()