+++ tests.py
@@ -14,6 +14,7 @@
 
 import os
 import sys
+import time
 from stat import S_ISDIR, S_ISREG, ST_MODE
 import unittest
 import subprocess
@@ -22,6 +23,12 @@ import signal
 
 import mocks
 import tcollector
+import json
+import threading
+try:
+    import flask
+except ImportError:
+    flask = None
 
 PY3 = sys.version_info[0] > 2
 
@@ -32,6 +39,35 @@ def return_none(x):
 
 def always_true():
     return True
+
+
+class ReaderThreadTests(unittest.TestCase):
+
+    def test_bad_float(self):
+        """Values that aren't ints/floats aren't sent to OpenTSDB.
+
+        This can happen if a specific collector is buggy.
+        """
+        thread = tcollector.ReaderThread(1, 10, True)
+        collector = tcollector.Collector("c", 1, "c")
+        for line in ["xxx", "mymetric 123 True a=b"]:
+            thread.process_line(collector, line)
+        self.assertEqual(thread.readerq.qsize(), 0)
+        self.assertEqual(collector.lines_received, 2)
+        self.assertEqual(collector.lines_invalid, 2)
+
+    def test_ok_lines(self):
+        """Good lines are passed on to OpenTSDB."""
+        thread = tcollector.ReaderThread(1, 10, True)
+        collector = tcollector.Collector("c", 1, "c")
+        for line in ["mymetric 123.24 12 a=b",
+                     "mymetric 124 12.7 a=b",
+                     "mymetric 125 12.7"]:
+            thread.process_line(collector, line)
+            self.assertEqual(thread.readerq.qsize(), 1, line)
+            self.assertEqual(thread.readerq.get(), line)
+        self.assertEqual(collector.lines_received, 3)
+        self.assertEqual(collector.lines_invalid, 0)
 
 
 class CollectorsTests(unittest.TestCase):
@@ -62,6 +98,106 @@ class CollectorsTests(unittest.TestCase):
             "/collectors/0"
         check_access_rights(collectors_path)
 
+    def test_json(self):
+        """A collector can be serialized to JSON."""
+        collector = tcollector.Collector("myname", 17, "myname.py", mtime=23, lastspawn=15)
+        collector.nextkill += 8
+        collector.killstate += 2
+        collector.lines_sent += 10
+        collector.lines_received += 65
+        collector.lines_invalid += 7
+        self.assertEqual(collector.to_json(),
+                         {"name": "myname",
+                          "mtime": 23,
+                          "lastspawn": 15,
+                          "killstate": 2,
+                          "nextkill": 8,
+                          "lines_sent": 10,
+                          "lines_received": 65,
+                          "lines_invalid": 7,
+                          "last_datapoint": collector.last_datapoint,
+                          "dead": False})
+
+
+class StatusServerTests(unittest.TestCase):
+    """Tests for StatusServer."""
+
+    def test_endtoend(self):
+        """We can get JSON status of collectors from StatusServer."""
+        collectors = {
+            "a": tcollector.Collector("mycollector", 5, "a.py"),
+            "b": tcollector.Collector("second", 3, "b.py"),
+        }
+        server = tcollector.StatusServer("127.0.0.1", 32025, collectors)
+        # runs in background until test suite exits :( but it works.
+        thread = threading.Thread(target=server.serve_forever)
+        thread.setDaemon(True)
+        thread.start()
+
+        r = tcollector.urlopen("http://127.0.0.1:32025").read()
+        self.assertEqual(json.loads(r), [c.to_json() for c in collectors.values()])
+
+
+@unittest.skipUnless(flask, "Flask not installed")
+class SenderThreadHTTPTests(unittest.TestCase):
+    """Tests for HTTP sending."""
+
+    def run_fake_opentsdb(self, response_code):
+        env = os.environ.copy()
+        env["FLASK_APP"] = "fake_opentsdb.py"
+        env["FAKE_OPENTSDB_RESPONSE"] = str(response_code)
+        flask = subprocess.Popen(["flask", "run", "--port", "4242"], env=env)
+        time.sleep(1)  # wait for it to start
+        self.addCleanup(flask.terminate)
+
+    def send_query_with_response_code(self, response_code):
+        """
+        Send a HTTP query using SenderThread, with server that returns given
+        response code.
+        """
+        self.run_fake_opentsdb(response_code)
+        reader = tcollector.ReaderThread(1, 10, True)
+        sender = tcollector.SenderThread(
+            reader, False, [("localhost", 4242)], False, {},
+            http=True, http_api_path="api/put"
+        )
+        sender.sendq.append("mymetric 123 12 a=b")
+        sender.send_data()
+        return sender
+
+    def test_normal(self):
+        """If response is OK, sendq is cleared."""
+        sender = self.send_query_with_response_code(204)
+        self.assertEqual(len(sender.sendq), 0)
+
+    def test_error(self):
+        """
+        If response is unexpected, e.g. 500 error, sendq is not cleared so we
+        can retry.
+        """
+        sender = self.send_query_with_response_code(500)
+        self.assertEqual(len(sender.sendq), 1)
+
+    def test_bad_messages(self):
+        """
+        If response is 400, sendq is cleared since there's no point retrying
+        bad messages.
+        """
+        sender = self.send_query_with_response_code(400)
+        self.assertEqual(len(sender.sendq), 0)
+
+class NamespacePrefixTests(unittest.TestCase):
+    """Tests for metric namespace prefix."""
+
+    def test_prefix_added(self):
+        """Namespace prefix gets added to metrics as they are read."""
+        thread = tcollector.ReaderThread(1, 10, True, "my.namespace.")
+        collector = tcollector.Collector("c", 1, "c")
+        line = "mymetric 123 12 a=b"
+        thread.process_line(collector, line)
+        self.assertEqual(thread.readerq.get(), "my.namespace." + line)
+        self.assertEqual(collector.lines_received, 1)
+        self.assertEqual(collector.lines_invalid, 0)
 
 class TSDBlacklistingTests(unittest.TestCase):
     """
@@ -125,7 +261,7 @@ class UDPCollectorTests(unittest.TestCase):
 
     def setUp(self):
         if ('udp_bridge.py' not in tcollector.COLLECTORS): # pylint: disable=maybe-no-member
-            return
+            raise unittest.SkipTest("udp_bridge unavailable")
 
         self.saved_exit = sys.exit
         self.saved_stderr = sys.stderr
@@ -348,6 +484,8 @@ class CollectorSanityCheckTests(unittest.TestCase):
 
 
 if __name__ == '__main__':
+    import logging
+    logging.basicConfig()
     cdir = os.path.join(os.path.dirname(os.path.realpath(sys.argv[0])),
                         'collectors')
     tcollector.setup_python_path(cdir) # pylint: disable=maybe-no-member